Move dlaran function
